Added accelerometer-based head-leveling reflex (and getting DataStreamCore to step inputs before and outputs after, a bit messy, could do with refactoring). Also added Compass sensor but couldn't make much sense of its behavior, will investigate when using second NXT.

+++ NXTAPE/src/org/opendrawer/nxtape/NXTAccelerometer.java
@@ -12,7 +12,7 @@ public class NXTAccelerometer implements DataProvider {
 			DataProvider.INPUT, DataProvider.INPUT };
 	private boolean inhibited = false;
 	private double values[];
-	private double maxExpectedValue = 256; // Arbitrary
+	private float maxExpectedValue = 256; // Arbitrary
 
 	public NXTAccelerometer(AccelHTSensor accelerometer, String name) {
 		this.accelerometer = accelerometer;
